# Create `thread_wait.v`: test []thread f64{} and wait performance.