Computer Configure：
···
CPU: 	Intel(R) Core(TM) i5-9500 CPU @ 3.00GHz   3.00 GHz
RAM	8.00 GB
OS            Windows 10 build 21318
···
And output:
```
$ ./wait
Thread *3* is running... only time.wait(1000ms)
Thread *2* is running... only time.wait(1000ms)
Thread *1* is running... only time.wait(1000ms)
>>> Thread - Total time used: 1002ms.                                                    // 66.67% time saved.
        Calculate `1+2+3+...+1000000000` ...
>>> Thread Sum  - Time used:  268ms, sum = 5e+017.                          // 80% time saved.
>>> `1+2+..max` - Time used: 1226ms, Sum = 5e+017.
```