@StunxFS 2021-02-24
Please run vfmt to the file: `v fmt -w examples/concurrency/thread_wait.v` :)