@UweKrueger 2021-02-24
The code in its present form leads to wrong results - especially for odd numbers of CPUs (e.g. 7). If you increase the displayed precision (e.g. with `println('${s:.3f}')` you can see the differences.

I've inserted some comments in the code that fix this and lead to correct results in both cases.